Fix glitch after root node has been moved
It seems that if the root node was moved (such as when setRelativeOrigin is
called), the root must be included for updateAllTransforms to work correctly.
If the root node is not included, there is a visible "glitch" in the rendering.
For simplicity, we always include the root node.

--HG--
branch : v2-1
extra : rebase_source : 90aff9527ec122753bcfa1b0dccc31a412ad6758